+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -88,7 +88,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		validator.addValidator(validator -> modelItems.get().stream()
 						.filter(Objects::nonNull)
 						.forEach(validator::test));
-		comparator.addListener(this::sortItems);
+		comparator.addListener(this::sort);
 		includeNull.addConsumer(value -> {
 			if (value && !modelItems.visible.items.contains(null)) {
 				modelItems.visible.items.add(0, null);
@@ -137,7 +137,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		if (modelItems.visiblePredicate.isNull() || modelItems.visiblePredicate.get().test(item)) {
 			if (!modelItems.visible.items.contains(item)) {
 				modelItems.visible.items.add(item);
-				sortItems();
+				sort();
 			}
 		}
 		else if (!modelItems.filtered.items.contains(item)) {
@@ -219,7 +219,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 	}
 
 	@Override
-	public final void sortItems() {
+	public final void sort() {
 		if (comparator.isNotNull() && !modelItems.visible.items.isEmpty()) {
 			if (includeNull.get()) {
 				modelItems.visible.items.remove(0);
@@ -410,7 +410,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 					}
 				}
 			}
-			sortItems();
+			sort();
 			if (selectionModel.selected.item != null && visible.items.contains(selectionModel.selected.item)) {
 				//update the selected item since the underlying data could have changed
 				selectionModel.selected.item = visible.items.get(visible.items.indexOf(selectionModel.selected.item));